% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{add_two}
\alias{add_two}
\title{Add Two Numbers}
\usage{
add_two(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of x and y.
}
\description{
This function adds two numbers together.
}
\examples{
add_two(2, 3)
}
