% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hema_Phylum_Result-data.R
\docType{data}
\name{Hema_Phylum_Result}
\alias{Hema_Phylum_Result}
\title{Example Output from ZIMMA}
\format{
A list with elements including:
\describe{
\item{Posterior Samples}{Posterior Samples}
\item{Phyloseq object}{Phyloseq object used to generate the results}
}
}
\usage{
data(Hema_Phylum_Result)
}
\description{
This dataset contains simulated posterior samples from a ZIMMA analysis,
including coefficient estimates and inclusion indicators for mediators.
}
\examples{
data(Hema_Phylum_Result)
str(Hema_Phylum_Result)
}
\keyword{datasets}
