% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{ZIMMA.summary}
\alias{ZIMMA.summary}
\title{Summarize ZIMMA Posterior Output}
\usage{
ZIMMA.summary(ZIMMA.pos = posterior_list)
}
\arguments{
\item{ZIMMA.pos}{The output from ZIMMA.}
}
\value{
The parameter's posterior inclusion probability, mean, median, highest posterior density.
}
\description{
This function summarizes the posterior results from ZIMMA.
}
\examples{
ZIMMA.summary(posterior_list)
}
